pyproject.toml
[project]
name = "small-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.5.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.11",
]
